requirements.txt
tqdm
python-dotenv
telethon
rarfile
sqlalchemy
pyzipper
